+++ src/Telegram.Bot/Types/CallbackQuery.cs
@@ -1,4 +1,5 @@
 using Newtonsoft.Json;
+using Telegram.Bot.Types.InlineKeyboardButtons;
 
 namespace Telegram.Bot.Types
 {
