+++ src/Telegram.Bot/Types/InlineKeyboardButtons/InlineKeyboardSwitchInlineQueryButton.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace Telegram.Bot.Types
+namespace Telegram.Bot.Types.InlineKeyboardButtons
 {
     /// <summary>
     /// This object represents one button of an inline keyboard that asks the user to switch to a chat
